Refactor: User Controller DTO 삭제 및 변경
