+++ src/main/java/com/sprint/mission/discodeit/controller/dto/IdResponse.java
@@ -1,0 +1,9 @@
+package com.sprint.mission.discodeit.controller.dto;
+
+import java.util.UUID;
+
+public record IdResponse(
+        boolean success,
+        UUID id
+) {
+}
