--- src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserIdResponse.java
@@ -1,9 +1,0 @@
-package com.sprint.mission.discodeit.controller.dto.user;
-
-import java.util.UUID;
-
-public record UserIdResponse(
-        boolean success,
-        UUID id
-) {
-}
