--- src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserStatusUpdateDataRequest.java
@@ -1,6 +1,0 @@
-package com.sprint.mission.discodeit.controller.dto.user;
-
-public record UserStatusUpdateDataRequest(
-        String status
-) {
-}
