--- src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserUpdateDataRequest.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.controller.dto.user;
-
-public record UserUpdateDataRequest(
-        String newUsername,
-        String newEmail,
-        String newPassword
-) {
-}
