--- src/main/java/com/sprint/mission/discodeit/entity/BinaryContentType.java
@@ -1,6 +1,0 @@
-package com.sprint.mission.discodeit.entity;
-
-public enum BinaryContentType {
-    USER_PROFILE,
-    MESSAGE_ATTACHMENT,
-}
